+++ src/nexgen/tools/Metafile.py
@@ -63,12 +63,19 @@ class DectrisMetafile(Metafile):
 
     @cached_property
     def hasConfig(self) -> bool:
-        if "config" in self._handle.keys():
+        if "_dectris" in self._handle.keys():
             return True
         return False
 
     def read_config_dset(self) -> Dict:
-        config = eval(self._handle["config"][()])
+        config = {}
+        for k, v in self._handle["_dectris"].items():
+            v = v[()]
+            if len(v) == 1:
+                v = v[0]
+                if isinstance(v, bytes):
+                    v = v.decode()
+            config[k] = v
         return config
 
     def get_number_of_images(self) -> int: